fix openseadragon svg - rename file; to check if viewer has been loaded
